cleanup path/bezier to prevent import triangle
